# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/documents/from-artifact": {
      "post": {
        "tags": [
          "conversion"
        ],
        "summary": "A021:add document",
        "description": "成果物をもとにドキュメントを新規追加する。\n\n指定されたアーティファクトIDとバージョンに基づいて成果物を取得し、\n新しいドキュメントIDを生成してドキュメントとして登録します。\n\n処理フロー:\n1. 成果物情報を取得\n2. 成果物IDを新しいドキュメントIDに変更\n3. S3に成果物を保存\n4. 成果物を新しいドキュメントとして登録（DB + MeilSearch）\n\n\nArgs:\n    request (AddDocumentRequest): 成果物新規追加リクエストデータ\n    user (User): 認証されたユーザー情報\n    conversion_service (ConversionService): ドキュメント操作サービス\n    s3_client (S3Client): S3クライアント\n\nReturns:\n    AddDocumentResponse: 新規追加処理の結果（成功/失敗、ドキュメントID、タイトル等）\n\nRaises:\n    HTTPException: 以下の場合にHTTPExceptionが発生します：\n        - 400: リクエストパラメータが不正な場合\n        - 404: 指定された成果物が存在しない場合\n        - 500: 内部サーバーエラーが発生した場合",
        "operationId": "A021",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddDocumentRequest",
                "description": "成果物新規追加リクエストデータ"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddDocumentResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string",
                      "description": "Error detail message"
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ]
                },
                "examples": {
                  "missing_token": {
                    "summary": "Missing token",
                    "value": {
                      "detail": "Authorization header missing"
                    }
                  },
                  "invalid_or_expired_token": {
                    "summary": "Invalid or expired token",
                    "value": {
                      "detail": "Invalid or expired token"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string",
                      "description": "Error detail message"
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ]
                },
                "examples": {
                  "internal_error": {
                    "value": {
                      "detail": "Internal Server Error"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string",
                      "description": "Error detail message"
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ]
                },
                "examples": {
                  "invalid_request": {
                    "value": {
                      "detail": "Invalid request parameters"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string",
                      "description": "Error detail message"
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ]
                },
                "examples": {
                  "not_found": {
                    "value": {
                      "detail": "Artifact not found"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "conversion"
        ],
        "summary": "A022:update document",
        "description": "成果物をもとにドキュメントを上書き保存する。\n\n指定されたアーティファクトIDとバージョンに基づいて成果物と関連ドキュメントのIDを取得し、\n既存のドキュメントを削除後、新しいドキュメントとして登録します。\n既存ドキュメントがコレクションに含まれている場合、コレクション内の参照も更新されます。\n\n処理フロー:\n1. 成果物情報を取得\n2. document_idで既存ドキュメントを検索\n3. S3に成果物を保存\n4. 成果物を新しいドキュメントとして登録（DB + MeilSearch）\n\n\nArgs:\n    request (UpdateDocumentRequest): 成果物更新リクエストデータ\n    user (User): 認証されたユーザー情報\n    update_service (UpdateService): 更新サービス\n    s3_client (S3Client): S3クライアント\n\nReturns:\n    UpdateDocumentResponse: 更新処理の結果（成功/失敗、ドキュメントID、タイトル等）\n\nRaises:\n    HTTPException: 以下の場合にHTTPExceptionが発生します：\n        - 400: リクエストパラメータが不正な場合\n        - 404: 指定された成果物が存在しない場合\n        - 500: 内部サーバーエラーが発生した場合",
        "operationId": "A022",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDocumentRequest",
                "description": "成果物更新リクエストデータ"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateDocumentResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string",
                      "description": "Error detail message"
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ]
                },
                "examples": {
                  "missing_token": {
                    "summary": "Missing token",
                    "value": {
                      "detail": "Authorization header missing"
                    }
                  },
                  "invalid_or_expired_token": {
                    "summary": "Invalid or expired token",
                    "value": {
                      "detail": "Invalid or expired token"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string",
                      "description": "Error detail message"
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ]
                },
                "examples": {
                  "internal_error": {
                    "value": {
                      "detail": "Internal Server Error"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string",
                      "description": "Error detail message"
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ]
                },
                "examples": {
                  "invalid_request": {
                    "value": {
                      "detail": "Invalid request parameters"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string",
                      "description": "Error detail message"
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ]
                },
                "examples": {
                  "not_found": {
                    "value": {
                      "detail": "Artifact not found"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AddDocumentRequest": {
        "properties": {
          "public_collection_id": {
            "type": "string",
            "maxLength": 11,
            "minLength": 11,
            "title": "Public Collection Id",
            "description": "公開コレクションの一意識別子（nanoid形式）",
            "examples": [
              "V1StGXR8_Z5"
            ]
          },
          "artifact_id": {
            "type": "string",
            "format": "uuid",
            "title": "Artifact Id",
            "description": "成果物の一意識別子（UUID形式）",
            "examples": [
              "3a92c5d1-6f3b-4b9a-bc2c-9f7d6e4f8e7a"
            ]
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version",
            "description": "成果物のバージョン番号（1以上）",
            "examples": [
              1
            ]
          }
        },
        "type": "object",
        "required": [
          "public_collection_id",
          "artifact_id",
          "version"
        ],
        "title": "AddDocumentRequest",
        "description": "成果物新規追加のリクエストモデル。\n\nAIによって生成された成果物を新しいドキュメントとして追加するための\nリクエストパラメータです。公開コレクションID、成果物の一意識別子とバージョンを指定します。\n\nAttributes:\n    public_collection_id: 公開コレクションの一意識別子（nanoid形式）\n    artifact_id: 成果物の一意識別子（UUID形式）\n    version: 成果物のバージョン番号（1以上の整数）"
      },
      "AddDocumentResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "新規追加処理の成功/失敗フラグ",
            "examples": [
              true
            ]
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "title": "Message",
            "description": "処理結果の詳細メッセージ",
            "examples": [
              "Document added successfully"
            ]
          },
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id",
            "description": "追加後のドキュメントの一意識別子（UUID形式）",
            "examples": [
              "260ea4dd-36a7-401f-b2a1-d6a28d97f140"
            ]
          },
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title",
            "description": "追加されたドキュメントのタイトル",
            "examples": [
              "プレゼンテーション資料"
            ]
          }
        },
        "type": "object",
        "required": [
          "success",
          "message",
          "document_id",
          "title"
        ],
        "title": "AddDocumentResponse",
        "description": "成果物新規追加のレスポンスモデル。\n\n成果物のドキュメント新規追加処理の結果を返却します。\n処理の成功/失敗、追加されたドキュメントの情報、処理結果メッセージを含みます。\n\nAttributes:\n    success: 新規追加処理が成功したかどうか\n    message: 処理結果の説明メッセージ\n    document_id: 追加後のドキュメントID（UUID形式）\n    title: 追加されたドキュメントのタイトル"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UpdateDocumentRequest": {
        "properties": {
          "public_collection_id": {
            "type": "string",
            "maxLength": 11,
            "minLength": 11,
            "title": "Public Collection Id",
            "description": "公開コレクションの一意識別子（nanoid形式）",
            "examples": [
              "V1StGXR8_Z5"
            ]
          },
          "artifact_id": {
            "type": "string",
            "format": "uuid",
            "title": "Artifact Id",
            "description": "成果物の一意識別子（UUID形式）",
            "examples": [
              "3a92c5d1-6f3b-4b9a-bc2c-9f7d6e4f8e7a"
            ]
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version",
            "description": "成果物のバージョン番号（1以上）",
            "examples": [
              1
            ]
          }
        },
        "type": "object",
        "required": [
          "public_collection_id",
          "artifact_id",
          "version"
        ],
        "title": "UpdateDocumentRequest",
        "description": "成果物更新のリクエストモデル。\n\nAIによって生成された成果物をドキュメントとして更新・登録するための\nリクエストパラメータです。公開コレクションID、成果物の一意識別子とバージョンを指定します。\n\nAttributes:\n    public_collection_id: 公開コレクションの一意識別子（nanoid形式）\n    artifact_id: 成果物の一意識別子（UUID形式）\n    version: 成果物のバージョン番号（1以上の整数）"
      },
      "UpdateDocumentResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "更新処理の成功/失敗フラグ",
            "examples": [
              true
            ]
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "title": "Message",
            "description": "処理結果の詳細メッセージ",
            "examples": [
              "Document updated successfully"
            ]
          },
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id",
            "description": "更新後のドキュメントの一意識別子（UUID形式）",
            "examples": [
              "260ea4dd-36a7-401f-b2a1-d6a28d97f140"
            ]
          },
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title",
            "description": "更新されたドキュメントのタイトル",
            "examples": [
              "プレゼンテーション資料"
            ]
          }
        },
        "type": "object",
        "required": [
          "success",
          "message",
          "document_id",
          "title"
        ],
        "title": "UpdateDocumentResponse",
        "description": "成果物更新のレスポンスモデル。\n\n成果物のドキュメント更新処理の結果を返却します。\n処理の成功/失敗、更新されたドキュメントの情報、処理結果メッセージを含みます。\n\nAttributes:\n    success: 更新処理が成功したかどうか\n    message: 処理結果の説明メッセージ\n    document_id: 更新後のドキュメントID（UUID形式）\n    title: 更新されたドキュメントのタイトル"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}